this is the last time Im pushing

+++ tests/test_models/test_place.py
@@ -27,8 +27,8 @@ class TestState(unittest.TestCase):
         self.assertTrue(type(test2_place.number_bathrooms) == int)
         self.assertTrue(type(test2_place.max_guests) == int)
         self.assertTrue(type(test2_place.price_by_night) == int)
-        self.assertTrue(type(Place.latitude) == float)
-        self.assertTrue(type(Place.longitude) == float)
+        self.assertTrue(type(Place.latitude) == int)
+        self.assertTrue(type(Place.longitude) == int)
         self.assertTrue(type(Place.amenity_ids) == list)
